proof of concept: only spawn one webserver
Previously this benchmark would spawn a new webserver for every
call of the benchmark function, which meant that it was always
testing an un-warmed-up process. It also executed a non-deterministic
number of requests per webserver, which would affect the results
by amortizing the warmup over a different number of requests.

With this hacky change, we spawn the webserver once, and then
begin the benchmarking process against it.