Remove the "mypyc" benchmark
The original intent was that we would try to ensure that we ran
the mypy benchmark with the "mypy-built-with-mypyc" build of mypy.

Something got lost at some point and this needs a little bit of work
to make sure that pyperf installs an appropriate build (such as what
run_mypy.sh does), but for now it's still an improvement to get rid
of the "mypyc" benchmark.

Fixes #5