Fix aiohttp benchmark on Python 3.11
This upgrades cython to a version that is compatible with Python 3.11, and then
uses `--no-binary` on `uvloop` and `yarl` to force a "re-Cythonization" of the
package.

I could understand if one doesn't want to merge this, since we would be testing
a local build of these packages rather than official upstream wheel. However,
this does get things working in the meantime.