Reduce the variability of the mypy benchmark
The first 2 runs of mypy read data from disk into a mypy-internal in-memory
cache.  This change attempts to reduce the impact of that I/O, which otherwise
causes the results to have an excessively wide range.

Additionally, since this makes the benchmark read as significantly "faster",
this renames the benchmark so it would be accidentally compared with baselines
against the old version.

There is more discussion here: https://github.com/faster-cpython/ideas/issues/543